+++ src/components/column.test.tsx
@@ -3,7 +3,7 @@ import boardState from '../util/board-state';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { createColumnJson, createKanbanJson, createTaskJson } from '../util/kanban-types';
-import { randomString, rightClick } from '../test-helpers';
+import { randomString, rightClick } from '../util/test-helpers';
 
 jest.mock('react-beautiful-dnd', () => {
     const dragDropElem = ({ children }: { children: Function }) =>
